povcalnet Facilitate interaction with the PovcalNet API

povcalnet Facilitate interaction with the PovcalNet API

povcalnet(country, povline = 1.9, year = "all", aggregate = FALSE,
  fill_gaps = FALSE, coverage = "all", ppp = NULL,
  url = "http://iresearch.worldbank.org", format = "csv")

Arguments

country
character: list of country iso3 code (accepts multiple) or `all`. Use povcalnet_info for full list of countries.

povline
numeric: poverty line (in 2011 PPP-adjusted USD) to calculate poverty measures

year
numeric: list of years, or `all`.

aggregate
logical: `TRUE` will return aggregate results, `FALSE` country-level results.

fill_gaps
logical: `TRUE` will interpolate / extrapolate values when surveys are not available for a specific year.

coverage
character: Can take one of three values: `national`, `urban`, `rural`

ppp
numeric: Optional - Allows the selection of custom PPP (Purchasing Power Parity) exchange rates

url
character: API root URL. For testing purposes only, should not be changed for 99 percent of users.

format
character: Response format to be requested from the API: `csv` or `json`

Value

data.frame

Examples

povcalnet(country = "ALB")

#> # A tibble: 5 x 31
#>   country_code country_name region_code coverage_type  year data_year data_type
#>   <chr>        <chr>        <chr>       <chr>         <dbl>     <dbl> <chr>    
#> 1 ALB          Albania      ECA         N              1996      1996 consumpt~
#> 2 ALB          Albania      ECA         N              2002      2002 consumpt~
#> 3 ALB          Albania      ECA         N              2005      2005 consumpt~
#> 4 ALB          Albania      ECA         N              2008      2008 consumpt~
#> 5 ALB          Albania      ECA         N              2012      2012 consumpt~
#> # ... with 24 more variables: is_interpolated <dbl>, use_microdata <dbl>,
#> #   ppp <dbl>, poverty_line <dbl>, mean <dbl>, headcount <dbl>,
#> #   poverty_gap <dbl>, poverty_gap_sq <dbl>, watts <dbl>, gini <dbl>,
#> #   median <dbl>, mld <dbl>, polarization <dbl>, population <dbl>,
#> #   decile1 <dbl>, decile2 <dbl>, decile3 <dbl>, decile4 <dbl>, decile5 <dbl>,
#> #   decile6 <dbl>, decile7 <dbl>, decile8 <dbl>, decile9 <dbl>, decile10 <dbl>